Rename p3_resolve_untyped -> p3_type_definitions
