+++ crates/solidity/outputs/cargo/crate/src/backend/passes/mod.rs
@@ -1,4 +1,4 @@
 pub mod p0_build_ast;
 pub mod p1_flatten_contracts;
 pub mod p2_collect_definitions;
-pub mod p3_resolve_untyped;
+pub mod p3_type_definitions;
