+++ crates/solidity/outputs/cargo/tests/src/backend/passes/pipeline.rs
@@ -82,7 +82,7 @@ fn test_backend_pipeline() -> Result<()> {
     let data = passes::p0_build_ast::run(unit);
     let data = passes::p1_flatten_contracts::run(data);
     let data = passes::p2_collect_definitions::run(data);
-    let data = passes::p3_resolve_untyped::run(data);
+    let data = passes::p3_type_definitions::run(data);
     assert_eq!(2, data.files.len());
 
     Ok(())
