+++ crates/solidity/outputs/cargo/tests/src/binder/runner.rs
@@ -6,7 +6,7 @@ use infra_utils::paths::PathExtensions;
 use metaslang_bindings::PathResolver;
 use semver::Version;
 use slang_solidity::backend::passes;
-use slang_solidity::backend::passes::p3_resolve_untyped::Output;
+use slang_solidity::backend::passes::p3_type_definitions::Output;
 use slang_solidity::compilation::{CompilationUnit, InternalCompilationBuilder};
 
 use super::renderer::binder_report;
@@ -85,5 +85,5 @@ fn build_binder(compilation_unit: CompilationUnit) -> Output {
     let data = passes::p0_build_ast::run(compilation_unit);
     let data = passes::p1_flatten_contracts::run(data);
     let data = passes::p2_collect_definitions::run(data);
-    passes::p3_resolve_untyped::run(data)
+    passes::p3_type_definitions::run(data)
 }
